Significant update - significant features implemented; lots of amendments
A lot of features are now present, but not finished. Tweets are responded to, and tweets from the user can alter the database and will soon affect the tweets sent back to the user.

Not every feature is complete or present. For example, investigating the user's surroundings, finding the key, or using it in room 10 to escape/'win' are not present. Placeholder text is still being used, and there are known bugs to be ironed out. Missing features are annotated throughout.